FIX BUG: Both the linear search and the binary search for fetching line numbers were wrong.
The linear search failed to account for the fact if there is no exact location match in the line number table, the line number will be given by entry immediately prior to the first entry whose location is past the current location.

The binary search was wrong about which endpoints to include when searching a half.